+++ template-java-core/src/test/java/com/retiman/template/lang/JavaLangTest.java
@@ -1,4 +1,4 @@
-package com.retiman.template.locale;
+package com.retiman.template.lang;
 
 import org.junit.jupiter.api.Test;
